requirements.txt
flask
google-cloud-logging
requests
